More Stream work.
Hidden wxMaxInputStream. wxImage and wxBitmap now test for filename, TStream and wxInputStream and pick the appropriate load routine. Fallback is to use the default wxWidgets loader.

Updated stream and splash samples to show use of incbin'ing.

Issue: On Linux at least, .xpm won't load via a stream for whatever reason - however, since it's unlikely that anyone will be using xpm in the real world, it's probably not a big deal.

git-svn-id: http://wxmax.googlecode.com/svn/trunk/wx.mod@91 971f8453-5c39-0410-a452-3998bf019487